requirements.txt
asgiref==3.8.1
Django==4.2.11
django-cors-headers==4.3.1
django-environ==0.11.2
django-storages==1.14.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pyflakes==2.2.0
PyJWT==2.8.0
pytz==2024.1
python-keycloak
sqlparse==0.4.2
psycopg2-binary>=2.8
pymongo
redis==4.5.1
celery==5.2.7
flower==1.2.0
django-celery-email==3.0.0
watchfiles==0.18.1